Removed strange contains check
The check

        if (!listeningVirtualAddresses.containsKey(vad))

checks if a VirtualAddress is a key contained in a HashMap<String, Set>.
This is strange since the VirtualAddress is not a String.
I think originally this test should have checked

        !listeningVirtualAddresses.get(topic).containsKey(vad)

But since the `get(topic)` is missing the check always returns `!false`
(since `vad.toString()` is not a topic) and it executes

    listeningVirtualAddresses.get(topic).add(vad);

which works because adding an element to a `Set` for a second time
results to the same set.

In addition to reduce the amount of code I stored the result of
`get(topic)` in a variable.